Each user is now connected to only one record of Questions and Averages. When the user registers a new account, a Questions and Averages record is made to the database with their default values of 0. This is so that when the user takes the test, it will load an instance of these records and update them in our database instead of creating another record with new results.

+++ owlwatch/app_owlwatch/models.py
@@ -17,8 +17,8 @@ class Averages(models.Model):
     """Averages model for storing average_intensity and average_frequency."""
 
     owner = models.ForeignKey(User, on_delete=models.CASCADE)
-    average_intensity = models.IntegerField(default=0)
-    average_frequency = models.IntegerField(default=0)
+    average_intensity = models.FloatField(default=0)
+    average_frequency = models.FloatField(default=0)
 
 
 class Questions(models.Model):
